requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
binance==0.3.6
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cron-descriptor==1.4.5
cryptography==44.0.2
cycler==0.12.1
dateparser==1.2.1
Django==5.2
django-celery-beat==2.8.0
django-timezone-field==7.1
djangorestframework==3.16.1
flower==2.0.1
fonttools==4.60.0
frozenlist==1.6.0
gunicorn==23.0.0
humanize==4.12.3
idna==3.10
kiwisolver==1.4.9
kombu==5.5.3
matplotlib==3.10.6
multidict==6.4.3
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
pyparsing==3.2.5
python-binance==1.0.28
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.6
pytz==2025.2
redis==6.0.0
regex==2024.11.6
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.0
